requirements.txt
discord.py
beautifulsoup4
py-cord